% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epical.R
\name{add_epi_week}
\alias{add_epi_week}
\title{Add Epi-weeks and Epi-years to a Data Frame}
\usage{
add_epi_week(data, date_col, ...)
}
\arguments{
\item{data}{A data frame}

\item{date_col}{Name of date column}

\item{...}{Additional parameters passed to \code{as.Date}}
}
\value{
A data frame
}
\description{
add_epi_week takes a data frame that already contains a date column and adds
columns for epi-week and epi-year.
}
\details{
If you do not provide the name of the date column in the supplied data frame,
the function looks for a column named "date". If it doesn't find one, it
stops with an error.
#'
}
\examples{
add_epi_week(data.frame(mydate = seq(as.Date("2014-12-26"),
                                 as.Date("2015-01-12"),
                                 by = "day")),
             "mydate")
}
\seealso{
Other epi.calendar.functions: \code{\link{epi_week_start}};
  \code{\link{epi_week}}; \code{\link{epi_year_start}}
}

