% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epiweek.R
\name{epi_week}
\alias{epi_week}
\title{Calculate the Epidemiological Week of Calendar Dates}
\usage{
epi_week(date)
}
\arguments{
\item{date}{A vector that can be coerced to a \code{Date}}
}
\value{
A numeric vector
}
\description{
Function to calculate the epidemiological week in which calendar
dates fall.
}
\details{
Epidemiological weeks, also called "epi-weeks" or "epi weeks", are a
standardized method of assigning days to a week of the year. Here we define
an epi-week as starting on a Sunday and ending on a Saturday, with the first
epi-week of the year ending on the first Saturday of the year that falls at
least four days into the year. In other words, the first epi-week always
begins on the Sunday that falls in the range of December 29 to January 4.

The function first converts \code{date} to a \code{Date} object using the
function \code{as.Date()}. I welcome suggestions for making this more robust.
}
\examples{
epi_week("2015-01-01")
epi_week(seq(as.Date("2014-12-26"), as.Date("2015-01-12"), by = "day"))
}
\seealso{
Other epi.calendar.functions: \code{\link{epi_week_start_date}};
  \code{\link{epi_year_start_date}}; \code{\link{epi_year}}
}

