% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epical.R
\name{epi_week_date}
\alias{epi_week_date}
\title{Calculate the Start Dates of Epidemiological Weeks}
\usage{
epi_week_date(epi_week, epi_year, offset = 0)
}
\arguments{
\item{epi_week}{An object that can be coerced to a \code{numeric}}

\item{epi_year}{An object that can be coerced to a \code{numeric}}

\item{offset}{A number of days to offset the returned dates by. The default
is 0 in which case the function returns the date of the first day of each
epi week. An offset of 6 would instead return the date of the last day of
each epi week.}
}
\value{
A \code{Date} vector
}
\description{
Function to calculate the start dates of epidemiological weeks.
}
\details{
Epidemiological weeks, also called "epi-weeks" or "epi weeks", are a
standardized method of assigning days to a week of the year. This function
determines the start date of a given set of epi-weeks identified by epi-week
number and year.

\code{epi_week} and \code{epi_year} are recycled if necessary.
}
\examples{
epi_week_date(1, 2010:2015)
epi_week_date(1:52, 2015)
epi_week_date(1:52, 2015, offset = 6) # return epi week end dates
epi_week_date(c(-1, 0, TRUE, 52, 53, 999, "a"), 2015) # NAs
}
\seealso{
Other epi.calendar.functions: \code{\link{add_epi_week}};
  \code{\link{epi_week}}; \code{\link{epi_year_start}}
}

