% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epical.R
\name{epi_week_start}
\alias{epi_week_start}
\title{Calculate the Start Dates of Epidemiological Weeks}
\usage{
epi_week_start(epi_week, epi_year)
}
\arguments{
\item{epi_week}{An object that can be coerced to a \code{numeric}}

\item{epi_year}{An object that can be coerced to a \code{numeric}}
}
\value{
A \code{Date} vector
}
\description{
Function to calculate the start dates of epidemiological weeks.
}
\details{
Epidemiological weeks, also called "epi-weeks" or "epi weeks", are a
standardized method of assigning days to a week of the year. This function
determines the start date of a given set of epi-weeks identified by epi-week
number and year.

\code{epi_week} and \code{epi_year} are recycled if necessary.
}
\examples{
epi_week_start(1, 2010:2015)
epi_week_start(1:52, 2015)
epi_week_start(c(-1, 0, TRUE, 52, 53, 999, "a"), 2015) # NAs
}
\seealso{
Other epi.calendar.functions: \code{\link{add_epi_week}};
  \code{\link{epi_week}}; \code{\link{epi_year_start}}
}

