% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epiweek.R
\name{epi_year}
\alias{epi_year}
\title{Calculate the Epidemiological Year of Calendar Dates}
\usage{
epi_year(date)
}
\arguments{
\item{date}{An object that can be converted to a \code{Date} by \code{as.Date()}}
}
\value{
A numeric vector
}
\description{
Function to calculate the epidemiological year in which calendar
dates fall.
}
\details{
Epidemiological weeks, also called "epi-weeks" or "epi weeks", are a
standardized method of assigning days to a week of the year. The first
epi-week always begins on the Sunday that falls in the range of December
29 to January 4. Thus dates in the first epidemiological week of the
year sometimes fall within the previous calendar year. This is a problem if
you want to group daily data into epi-weeks and years.

Enter the "epi-year", which starts on the first day of epidemiological week 1
and ends on the last day before the next epidemiological year starts.

The function first converts \code{x} to a \code{Date} object using the
function \code{as.Date()}. I welcome suggestions for making this more robust.
}
\examples{
epi_year(c("2015-01-03", "2015-01-04"))
epi_year(c("2013-12-28", "2013-12-29"))
}
\seealso{
Other epi.calendar.functions: \code{\link{epi_week_start_date}};
  \code{\link{epi_week}}; \code{\link{epi_year_start_date}}
}

