% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epical.R
\name{epi_year_start}
\alias{epi_year_start}
\title{#' Start Dates of Epidemiological Years}
\usage{
epi_year_start(year)
}
\arguments{
\item{year}{An object that can be coerced to a \code{numeric}}
}
\value{
A \code{Date} vector
}
\description{
Function to calculate the start dates of epidemiological years.
}
\details{
The first epi-week always begins on the Sunday that falls in the range of
December 29 to January 4. This function determines which date of the first
day of epi-week 1. Note that the year of the returned date may not be the
same as the year provided, in cases where epi-week 1 begins in December of
the previous year.

I'm not sure if it can handle NAs. Some improvements may be needed.
}
\examples{
epi_year_start(c(2014,2015))
}
\seealso{
Other epi.calendar.functions: \code{\link{add_epi_week}};
  \code{\link{epi_week_date}}; \code{\link{epi_week}}
}

