% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epical.R
\name{epi_year_start_date}
\alias{epi_year_start_date}
\title{#' Calculate the Start Dates of Epidemiological Years}
\usage{
epi_year_start_date(year)
}
\arguments{
\item{year}{An object that can be coerced to a \code{numeric}}
}
\value{
A \code{Date} vector
}
\description{
Function to calculate the start dates of epidemiological years.
}
\details{
Epidemiological weeks, also called "epi-weeks" or "epi weeks", are a
standardized method of assigning days to a week of the year. The first
epi-week always begins on the Sunday that falls in the range of December
29 to January 4. This function determines which date the first day of
epi-week 1 falls in.

I'm not sure if it can handle NAs. Some improvements may be needed.
}
\examples{
epi_year_start_date(c(2014,2015))
}
\seealso{
Other epi.calendar.functions: \code{\link{add_epi_week}};
  \code{\link{epi_week_start_date}}; \code{\link{epi_week}}
}

