% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epical.R
\docType{package}
\name{epical}
\alias{epical}
\alias{epical-package}
\title{Convert between dates and epidemiological weeks/years.}
\description{
epical allows you to calculate back and forth between epi-weeks and
calendar dates.
}
\details{
To determine which epi-week a calendar date falls in, use
\code{\link{epi_week}}.

To determine the start date of an epi-week or epi-year, use
\code{\link{epi_week_start_date}} or \code{\link{epi_week_start_year}}.
}

