% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epical.R
\name{epi_week}
\alias{epi_week}
\title{Date Conversion to Epidemiological Weeks}
\usage{
epi_week(x, system = "who", ...)
}
\arguments{
\item{x}{A Date vector, or another vector that can be converted to Date using
\code{as.Date}}

\item{system}{Either "who" or "cdc". WHO epidemiological weeks start on
Monday. CDC epidemiological weeks (MMWR weeks) start on Sunday}

\item{...}{Additional parameters passed to \code{as.Date}}
}
\value{
A data frame with columns labeled \code{epi_week} and
\code{epi_year}. The object also has class tbl_df and tbl for use with the
\code{dplyr} package.
}
\description{
Function to calculate the epidemiological weeks in which one or more calendar
dates fall.
}
\details{
The function first converts \code{date} to a \code{Date} object using the
function \code{as.Date()}. I welcome suggestions for making this more robust.
}
\examples{
epi_week("2015-01-01")
epi_week("2015-150", format = "\%Y-\%j")
epi_week(seq(as.Date("2014-12-26"), as.Date("2015-01-12"), by = "day"))
}
\seealso{
Other epi.calendar.functions: \code{\link{add_epi_week}};
  \code{\link{epi_week_date}}; \code{\link{epi_year_start}}
}

