% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epical.R
\name{epi_week}
\alias{epi_week}
\title{Calculate Epidemiological Weeks from Calendar Dates}
\usage{
epi_week(x, ...)
}
\arguments{
\item{x}{A Date vector, or another vector that can be converted to Date using
\code{as.Date}}

\item{...}{Additional parameters passed to \code{as.Date}}
}
\value{
A data frame with columns labeled \code{epi_week} and
\code{epi_year}. The object also has class tbl_df and tbl for use with the
\code{dplyr} package.
}
\description{
Function to calculate the epidemiological week and year in which calendar
dates fall.
}
\details{
Epidemiological weeks, also called "epi-weeks" or "epi weeks", are a
standardized method of assigning days to a week of the year. Epi-weeks are
used by the WHO, CDC, and many other health organizations.

Here we define an epi-week as starting on a Sunday and ending on a Saturday,
with the first epi-week of the year ending on the first Saturday of the year
that falls at least four days into the year. In other words, the first
epi-week always begins on the Sunday that falls in the range of December 29
to January 4.

Because some dates in epi-week 1 can fall within the previous
calendar year, and dates in epi-week 52/53 can fall within the next calendar
year, it is necessary to specify both the epi-week number and an
"epi-year" so that dates can be grouped correctly.

The function first converts \code{date} to a \code{Date} object using the
function \code{as.Date()}. I welcome suggestions for making this more robust.
}
\examples{
epi_week("2015-01-01")
epi_week("2015-150", format = "\%Y-\%j")
epi_week(seq(as.Date("2014-12-26"), as.Date("2015-01-12"), by = "day"))
}
\seealso{
Other epi.calendar.functions: \code{\link{add_epi_week}};
  \code{\link{epi_week_start_date}};
  \code{\link{epi_year_start_date}}
}

