% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epical.R
\name{epi_week_date}
\alias{epi_week_date}
\title{Date Conversion from Epidemiological Weeks}
\usage{
epi_week_date(epi_week, epi_year, offset = 0, system = "who")
}
\arguments{
\item{epi_week}{An object that can be coerced to a \code{numeric}}

\item{epi_year}{An object that can be coerced to a \code{numeric}}

\item{offset}{A number of days to offset the returned dates by. The default
is 0 in which case the function returns the date of the first day of each
epi week. An offset of 6 would instead return the date of the last day of
each epi week.}

\item{system}{Either "who" or "cdc". Defauly is "who".
See \code{\link{epi_week}} for details.}
}
\value{
A \code{Date} vector
}
\description{
Function to calculate dates from epidemiological weeks.
}
\details{
This function determines the date of the first day of each epi-week,
optionally adding an offset to calculate the date of the last day of the
week, or any other arbitrary amount.

\code{epi_week} and \code{epi_year} are recycled if necessary.
}
\examples{
epi_week_date(1, 2010:2015)
epi_week_date(1:52, 2015)
epi_week_date(1:52, 2015, offset = 6) # return epi week end dates
epi_week_date(c(-1, 0, TRUE, 52, 53, 999, "a", NA), 2015) # NAs
}
\seealso{
Other epi.calendar.functions: \code{\link{add_epi_week}};
  \code{\link{epi_week}}; \code{\link{epi_year_start}}
}

